feat: costs as int to avoid issues with floating point precision. cost factor for diagonal movement on grid. increase grid size in tests. github action to run automated tests

+++ graph/dijkstra.go
@@ -23,7 +23,7 @@ func (d *Dijkstra) CalculateCostPath(graph *Graph, start string, goal string) (C
 	d.Visited = map[string]string{
 		start: "",
 	}
-	d.Costs = map[string]float64{
+	d.Costs = map[string]int{
 		start: 0,
 	}
